requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
APScheduler==3.6.3
beautifulsoup4==4.9.0
blinker==1.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-WTF==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==2.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
pytz==2020.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.0
typing_extensions==4.12.2
tzlocal==2.0.0
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
